Force la taille du paramètre de génération de valeurs aléatoire à 32 bits pour éviter des valeurs différentes entre les architectures 32 et 64 bits.

+++ uno/core/types.cpp
@@ -1,7 +1,7 @@
 #include "types.h"
 
 /// Paramètre à utiliser pour la génération pseudo-aléatoire des nombres.
-static unsigned long int next = 1;
+static uint32_t next = 1;
 
 int my_rand(void)
 {
